openjdk-7: enable xawt
The build of xawt has been disabled as its build seemed to fail during a qemu
call to generate a file with information on variable sizes.

This patch enables xawt and fixes the qemu run.

GENERATING .../icedtea-2.3.12/build/openjdk.build/gensrc/sun/awt/X11/generator/sizes.32
/lib/ld-linux-armhf.so.3: No such file or directory

Despite the error this generated an empty sizes.32 file. On subsequent bitbake
runs the compilation did not rerun qemu and followup errors ended the bitbake.

linking the sizer.32 binary with -static fixed this.
As a safeguard an empty sizes.32 file is now deleted to make also subsequent
builds fail at the position of the inital error.