qemu binaries: remove linking with -static
For colibri-t20 a build of openjdk-7 fails with segmentation faults during
the qemu-arm runs.

My investigation showed that qemu-arm can not execute any binary which
is compiled with the compiler for Colibri-T20 and is linked with the -static
flag. Even a simple hello world program does not work. When executing that
hello world on the real hardware I also get a segmentation fault.

Removing the -static flag fixed this.

I do not think that the fix with removing '-static' should go into meta-java!
We rather should find why the compiler does not produce working binaries.